refactor - create ContributionsEpicCtas component
