+++ packages/modules/src/modules/epics/ContributionsLiveblogEpic.stories.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { ContributionsLiveblogEpic } from './ContributionsLiveblogEpic';
 import { Story, Meta } from '@storybook/react';
-import { EpicProps } from '@sdc/shared/types';
+import { EpicProps, SecondaryCtaType } from '@sdc/shared/types';
 import { props } from './utils/storybook';
-import { WithReminder } from '../banners/contributions/ContributionsBanner.stories';
 
 export default {
     component: ContributionsLiveblogEpic,
@@ -13,24 +12,43 @@ export default {
 
 const Template: Story<EpicProps> = (props: EpicProps) => <ContributionsLiveblogEpic {...props} />;
 
-export const Default = Template.bind({});
-
-export const WithoutSupportUrl = Template.bind({});
-WithoutSupportUrl.args = {
-    ...WithReminder.args,
-    variant: {
-        ...props.variant,
-        cta: {
-            baseUrl: 'https://theguardian.com',
-            text: 'The Guardian',
-        },
-    },
-};
-
 export const WhenOnDCR = Template.bind({});
 WhenOnDCR.args = {
     tracking: {
         ...props.tracking,
         clientName: 'dcr',
     },
+    variant: {
+        ...props.variant,
+        secondaryCta: undefined,
+    },
 };
+
+export const WithoutSupportUrl = Template.bind({});
+WithoutSupportUrl.args = {
+    variant: {
+        ...props.variant,
+        cta: {
+            baseUrl: 'https://theguardian.com',
+            text: 'The Guardian',
+        },
+        secondaryCta: undefined,
+    },
+};
+
+export const WithReminderCta = Template.bind({});
+WithReminderCta.args = {
+    variant: {
+        ...props.variant,
+        secondaryCta: {
+            type: SecondaryCtaType.ContributionsReminder,
+        },
+        showReminderFields: {
+            reminderCta: 'Remind me in December',
+            reminderPeriod: '2022-12-01',
+            reminderLabel: 'December',
+        },
+    },
+};
+
+export const WithBespokeCta = Template.bind({});
